refactor(app): extract tab content rendering into useTabContentRenderer hook
getTabLabel/renderTabContent/renderTabBar moved out of App.tsx into a
dedicated hook taking a single params object, mirroring the useProjectIO
and useFileSystemManager extraction pattern. Reduces App.tsx by ~264 lines.

Closes bmf-vangard-renpy-ide-4oq.
